requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
